+++ sentry/src/main/java/io/sentry/DataCategory.java
@@ -11,6 +11,7 @@ public enum DataCategory {
   Attachment("attachment"),
   LogItem("log_item"),
   LogByte("log_byte"),
+  TraceMetric("trace_metric"),
   Monitor("monitor"),
   Profile("profile"),
   ProfileChunkUi("profile_chunk_ui"),