+++ sentry/src/test/java/io/sentry/protocol/SentryItemTypeSerializationTest.kt
@@ -33,6 +33,7 @@ class SentryItemTypeSerializationTest {
     assertEquals(serialize(SentryItemType.CheckIn), json("check_in"))
     assertEquals(serialize(SentryItemType.Feedback), json("feedback"))
     assertEquals(serialize(SentryItemType.Span), json("span"))
+    assertEquals(serialize(SentryItemType.TraceMetric), json("trace_metric"))
   }
 
   @Test
@@ -51,6 +52,7 @@ class SentryItemTypeSerializationTest {
     assertEquals(deserialize(json("check_in")), SentryItemType.CheckIn)
     assertEquals(deserialize(json("feedback")), SentryItemType.Feedback)
     assertEquals(deserialize(json("span")), SentryItemType.Span)
+    assertEquals(deserialize(json("trace_metric"), SentryItemType.TraceMetric)
   }
 
   private fun json(type: String): String = "{\"type\":\"${type}\"}"